Update location.ts, update documentation, move updateUrlParameter to location
